refactor(workers): separate worker logic into their own services
