+++ src/services/question/worker/aiSuggestion.service.ts
@@ -1,0 +1,25 @@
+import generateQuestionSuggestionService from "../ai/generateQuestionSuggestion.service.js";
+
+import { getRedisCacheClient } from "../../../config/redis.config.js";
+
+type ProcessAiSuggestionJobData = {
+  userId: string;
+  questionId: string;
+  version: number;
+};
+
+const processAiSuggestionJob = async ({
+  userId,
+  questionId,
+  version,
+}: ProcessAiSuggestionJobData) => {
+  try {
+    await generateQuestionSuggestionService({ userId, questionId, version });
+  } finally {
+    await getRedisCacheClient().del(
+      `aiSuggestion:pending:${userId}:${questionId}:${version}`,
+    );
+  }
+};
+
+export default processAiSuggestionJob;
