+++ src/services/question/worker/contentPipelineRouter.service.ts
@@ -1,0 +1,27 @@
+import answerPipelineRouterService from "../pipelineRouter/answerPipelineRouter.service.js";
+import aiAnswerFeedbackPipelineRouterService from "../pipelineRouter/aiAnswerFeedbackPipelineRouter.service.js";
+import questionPipelineRouterService from "../pipelineRouter/questionPipelineRouter.service.js";
+import replyPipelineRouterService from "../pipelineRouter/replyPipelineRouter.service.js";
+
+const processContentPipelineRouterJob = async (
+  contentType: "QUESTION" | "ANSWER" | "REPLY" | "AI_ANSWER_FEEDBACK",
+  contentId: string,
+  version?: number,
+) => {
+  switch (contentType) {
+    case "QUESTION":
+      await questionPipelineRouterService(contentId, version as number);
+      break;
+    case "ANSWER":
+      await answerPipelineRouterService(contentId);
+      break;
+    case "REPLY":
+      await replyPipelineRouterService(contentId);
+      break;
+    case "AI_ANSWER_FEEDBACK":
+      await aiAnswerFeedbackPipelineRouterService(contentId);
+      break;
+  }
+};
+
+export default processContentPipelineRouterJob;
