+++ src/workers/media/imageDeletion.worker.ts
@@ -1,22 +1,13 @@
 import { Worker } from "bullmq";
-import { redisMessagingClientConnection } from "../../config/redis.config.js";
 
-import deleteSingleImage from "../../services/media/deleteSingleImage.service.js";
-import deleteImagesFromBody from "../../services/media/deleteImageFromBody.service.js";
+import processImageDeletionJob from "../../services/media/worker/imageDeletion.service.js";
+
+import { redisMessagingClientConnection } from "../../config/redis.config.js";
 
 const worker = new Worker(
   "imageDeletionQueue",
   async (job) => {
-    switch (job.name) {
-      case "DELETE_SINGLE":
-        await deleteSingleImage(job.data);
-        break;
-      case "DELETE_FROM_BODY":
-        await deleteImagesFromBody(job.data);
-        break;
-      default:
-        throw new Error("Invalid job type");
-    }
+    await processImageDeletionJob(job.name, job.data);
   },
   {
     connection: redisMessagingClientConnection,
