+++ src/workers/user/userInterest.worker.ts
@@ -1,76 +1,12 @@
 import { Worker } from "bullmq";
 import { fileURLToPath } from "node:url";
 
-import { redisMessagingClientConnection } from "../../config/redis.config.js";
+import processUserInterestJob from "../../services/user/worker/userInterest.service.js";
+
 import connectMongoDB from "../../config/mongodb.config.js";
-
-import UserInterest from "../../models/userInterest.model.js";
-
-import type { UserInterestAction } from "../../utils/question/queueUserInterest.util.js";
+import { redisMessagingClientConnection } from "../../config/redis.config.js";
 
 const workerFilePath = fileURLToPath(import.meta.url);
-
-const actionScores = {
-  VIEW: 1,
-  UPVOTE: 3,
-  ANSWER: 5,
-} as const;
-
-function isUserInterestAction(action: string): action is UserInterestAction {
-  return action in actionScores;
-}
-
-async function applyInterestScore(userId: string, tag: string, score: number) {
-  await UserInterest.updateOne(
-    { userId },
-    [
-      {
-        $set: {
-          userId,
-          interests: {
-            $let: {
-              vars: {
-                existingInterests: { $ifNull: ["$interests", []] },
-                existingTags: {
-                  $map: {
-                    input: { $ifNull: ["$interests", []] },
-                    as: "interest",
-                    in: "$$interest.tag",
-                  },
-                },
-              },
-              in: {
-                $cond: [
-                  { $in: [tag, "$$existingTags"] },
-                  {
-                    $map: {
-                      input: "$$existingInterests",
-                      as: "interest",
-                      in: {
-                        $cond: [
-                          { $eq: ["$$interest.tag", tag] },
-                          {
-                            tag: "$$interest.tag",
-                            score: { $add: ["$$interest.score", score] },
-                          },
-                          "$$interest",
-                        ],
-                      },
-                    },
-                  },
-                  {
-                    $concatArrays: ["$$existingInterests", [{ tag, score }]],
-                  },
-                ],
-              },
-            },
-          },
-        },
-      },
-    ],
-    { upsert: true },
-  );
-}
 
 async function startUserInterestWorker() {
   await connectMongoDB(process.env.MONGO_URI as string);
@@ -78,21 +14,7 @@ async function startUserInterestWorker() {
   const worker = new Worker(
     "userInterestQueue",
     async (job) => {
-      const action = job.name;
-
-      if (!isUserInterestAction(action))
-        throw new Error(`Unsupported user interest action: ${action}`);
-
-      const { userId, tags } = job.data as {
-        userId: string;
-        tags: string[];
-      };
-      const score = actionScores[action];
-      const uniqueTags = [...new Set(tags)];
-
-      for (const tag of uniqueTags) {
-        await applyInterestScore(userId, tag, score);
-      }
+      await processUserInterestJob(job.name, job.data);
     },
     {
       connection: redisMessagingClientConnection,
